Moved ConceptPreviewType to interfaces
